DVX-TK-021 Fix Room database factory: move expect/actual to composeApp
Move createDividendDatabaseBuilder() expect/actual from :component:dividend
to :composeApp following the CurrencyModule pattern. The Android actual
requires Context via Koin which must live in the app module.

- Remove DividendDatabaseFactory.kt from component:dividend
- Add expect fun createDividendDatabaseBuilder() to composeApp/commonMain
- Add actual implementations to composeApp/android|jvm|iosMain
- Update buildDividendLocalDataSource() to accept DividendDao parameter
- Update DividendModule.kt to wire DividendDatabase -> dao -> local source
- Expose room-runtime as api in component:dividend for RoomDatabase type
