+++ composeApp/src/commonMain/kotlin/com/akole/dividox/di/DividendModule.kt
@@ -1,7 +1,9 @@
 package com.akole.dividox.di
 
+import com.akole.dividox.component.auth.domain.usecase.GetCurrentUserIdUseCase
 import com.akole.dividox.component.dividend.data.datasource.DividendLocalDataSource
 import com.akole.dividox.component.dividend.data.datasource.DividendRemoteDataSource
+import com.akole.dividox.component.dividend.data.db.DividendDatabase
 import com.akole.dividox.component.dividend.data.db.buildDividendLocalDataSource
 import com.akole.dividox.component.dividend.data.repository.DividendRepositoryImpl
 import com.akole.dividox.component.dividend.domain.repository.DividendRepository
@@ -10,22 +12,25 @@ import com.akole.dividox.component.dividend.domain.usecase.GetDividendHistoryUse
 import com.akole.dividox.component.dividend.domain.usecase.GetLifetimeDividendsUseCase
 import com.akole.dividox.component.dividend.domain.usecase.GetUpcomingPaymentsUseCase
 import com.akole.dividox.component.dividend.domain.usecase.GetYtdDividendsUseCase
-import com.akole.dividox.component.auth.domain.usecase.GetCurrentUserIdUseCase
 import org.koin.dsl.module
 
 /**
  * Koin module for the dividend component.
  *
  * Wiring:
- * 1. [DividendDatabase] — created via platform-specific [buildDividendDatabase].
- * 2. [DividendLocalDataSource] — wraps the Room DAO.
+ * 1. [DividendDatabase] — created via platform-specific [createDividendDatabaseBuilder].
+ * 2. [DividendLocalDataSource] — wraps the Room DAO via [buildDividendLocalDataSource].
  * 3. [DividendRemoteDataSource] — wraps Firestore, scoped to the current user.
  * 4. [DividendRepository] — [DividendRepositoryImpl] bridging local + remote.
  * 5. Use cases — one factory per use case.
  */
 val dividendModule = module {
+    single<DividendDatabase> {
+        createDividendDatabaseBuilder().build()
+    }
+
     single<DividendLocalDataSource> {
-        buildDividendLocalDataSource()
+        buildDividendLocalDataSource(get<DividendDatabase>().dividendDao())
     }
 
     single {
